requirements.txt
requests_html==0.10.0
pandas==1.1.3
requests==2.24.0
numpy==1.19.2
bdshare==0.1.4
beautifulsoup4==4.9.3
python_dateutil==2.8.1
